--- packages/server/src/app/api/content/news/news.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { NewsController } from './news.controller';
-
-describe('News Controller', () => {
-    let controller: NewsController;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            controllers: [NewsController],
-        }).compile();
-
-        controller = module.get<NewsController>(NewsController);
-    });
-
-    it('should be defined', () => {
-        expect(controller).toBeDefined();
-    });
-});
